pyproject.toml
[project]
name = "lightrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=14.3.0",
    "aiohttp>=3.11.18",
    "configparser>=7.2.0",
    "future>=1.0.0",
    "markdown>=3.8",
    "milvus>=2.3.9",
    "nano-vectordb>=0.0.4.3",
    "neo4j>=5.28.1",
    "nest-asyncio>=1.6.0",
    "pandas>=2.0.0",
    "pipmaster>=0.7.2",
    "pydantic>=2.11.4",
    "pymilvus>=2.5.8",
    "python-dotenv>=1.1.0",
    "pyuca>=1.2",
    "setuptools>=80.4.0",
    "tenacity>=9.1.2",
    "tiktoken>=0.9.0",
    "xlsxwriter>=3.1.0",
]
